+++ probdiffeq/statespace/dense/_conds.py
@@ -4,8 +4,8 @@
 import jax
 
 from probdiffeq import _sqrt_util
-from probdiffeq.ssm import _collections
-from probdiffeq.ssm.dense import _vars
+from probdiffeq.statespace import _collections
+from probdiffeq.statespace.dense import _vars
 
 
 @jax.tree_util.register_pytree_node_class