+++ probdiffeq/statespace/scalar/_vars.py
@@ -4,8 +4,8 @@ import jax
 import jax.numpy as jnp
 
 from probdiffeq import _sqrt_util
-from probdiffeq.ssm import _collections
-from probdiffeq.ssm.scalar import _conds
+from probdiffeq.statespace import _collections
+from probdiffeq.statespace.scalar import _conds
 
 
 @jax.tree_util.register_pytree_node_class