+++ probdiffeq/statespace/scalar/extra.py
@@ -1,18 +1,18 @@
-"""Extrapolations."""
+"""Implementations for scalar initial value problems."""
 
 import jax
 import jax.numpy as jnp
 
 from probdiffeq import _sqrt_util
-from probdiffeq.ssm import _collections, _ibm_util
-from probdiffeq.ssm.iso import _conds, _vars
+from probdiffeq.statespace import _collections, _ibm_util
+from probdiffeq.statespace.scalar import _conds, _vars
 
 
-def ibm_iso(num_derivatives):
+def ibm_scalar(num_derivatives):
     a, q_sqrtm = _ibm_util.system_matrices_1d(num_derivatives=num_derivatives)
     _tmp = _ibm_util.preconditioner_prepare(num_derivatives=num_derivatives)
     scales, powers = _tmp
-    return _IsoIBM(
+    return _IBM(
         a=a,
         q_sqrtm_lower=q_sqrtm,
         preconditioner_scales=scales,
@@ -21,7 +21,7 @@ def ibm_iso(num_derivatives):
 
 
 @jax.tree_util.register_pytree_node_class
-class _IsoIBM(_collections.AbstractExtrapolation):
+class _IBM(_collections.AbstractExtrapolation):
     def __init__(self, a, q_sqrtm_lower, preconditioner_scales, preconditioner_powers):
         self.a = a
         self.q_sqrtm_lower = q_sqrtm_lower
@@ -31,7 +31,10 @@ class _IsoIBM(_collections.AbstractExtrapolation):
 
     def __repr__(self):
         name = self.__class__.__name__
-        return f"{name}(a={self.a}, q_sqrtm_lower={self.q_sqrtm_lower})"
+        args1 = f"a={self.a}, q={self.q_sqrtm_lower}"
+        args2 = f"preconditioner_scales={self.preconditioner_scales}"
+        args3 = f"preconditioner_powers={self.preconditioner_powers}"
+        return f"{name}({args1}, {args2}, {args3})"
 
     def tree_flatten(self):
         children = (
@@ -62,39 +65,27 @@ class _IsoIBM(_collections.AbstractExtrapolation):
             msg1 = "The number of Taylor coefficients does not match "
             msg2 = "the number of derivatives in the implementation."
             raise ValueError(msg1 + msg2)
-        m0_corrected = jnp.stack(taylor_coefficients)
+
+        m0_matrix = jnp.stack(taylor_coefficients)
+        m0_corrected = jnp.reshape(m0_matrix, (-1,), order="F")
         c_sqrtm0_corrected = jnp.zeros_like(self.q_sqrtm_lower)
-        rv = _vars.IsoNormalHiddenState(
+
+        rv = _vars.NormalHiddenState(
             mean=m0_corrected, cov_sqrtm_lower=c_sqrtm0_corrected
         )
-        return _vars.IsoStateSpaceVar(rv, cache=None)
-
-    # todo: why does this method have the same name as the above?
-    def init_ssv(self, ode_shape):
-        assert len(ode_shape) == 1
-        (d,) = ode_shape
-        m0 = jnp.zeros((self.num_derivatives + 1, d))
-        c0 = jnp.eye(self.num_derivatives + 1)
-        rv = _vars.IsoNormalHiddenState(m0, c0)
-        return _vars.IsoStateSpaceVar(rv, cache=None)
+        return _vars.StateSpaceVar(rv, cache=None)
 
     def init_error_estimate(self):
-        return jnp.zeros(())  # the initialisation is error-free
+        return jnp.zeros(())
 
-    def init_output_scale(self, output_scale):
-        return output_scale
-
-    def begin_extrapolation(
-        self, s0: _vars.IsoStateSpaceVar, /, dt
-    ) -> _vars.IsoStateSpaceVar:
+    def begin_extrapolation(self, s0, /, dt):
         p, p_inv = self._assemble_preconditioner(dt=dt)
-        m0_p = p_inv[:, None] * s0.hidden_state.mean
+        m0_p = p_inv * s0.hidden_state.mean
         m_ext_p = self.a @ m0_p
-        m_ext = p[:, None] * m_ext_p
+        m_ext = p * m_ext_p
         q_sqrtm = p[:, None] * self.q_sqrtm_lower
-
-        ext = _vars.IsoNormalHiddenState(m_ext, q_sqrtm)
-        return _vars.IsoStateSpaceVar(ext, cache=(m_ext_p, m0_p, p, p_inv))
+        extrapolated = _vars.NormalHiddenState(m_ext, q_sqrtm)
+        return _vars.StateSpaceVar(extrapolated, cache=(m_ext_p, m0_p, p, p_inv))
 
     def _assemble_preconditioner(self, dt):
         return _ibm_util.preconditioner_diagonal(
@@ -106,17 +97,16 @@ class _IsoIBM(_collections.AbstractExtrapolation):
     ):
         _, _, p, p_inv = output_begin.cache
         m_ext = output_begin.hidden_state.mean
-
-        l0_p = p_inv[:, None] * s0.hidden_state.cov_sqrtm_lower
         l_ext_p = _sqrt_util.sum_of_sqrtm_factors(
             R_stack=(
-                (self.a @ l0_p).T,
+                (self.a @ (p_inv[:, None] * s0.hidden_state.cov_sqrtm_lower)).T,
                 (output_scale * self.q_sqrtm_lower).T,
             )
         ).T
         l_ext = p[:, None] * l_ext_p
-        rv = _vars.IsoNormalHiddenState(m_ext, l_ext)
-        return _vars.IsoStateSpaceVar(rv, cache=None)
+
+        rv = _vars.NormalHiddenState(mean=m_ext, cov_sqrtm_lower=l_ext)
+        return _vars.StateSpaceVar(rv, cache=None)
 
     def complete_extrapolation_with_reversal(self, output_begin, /, s0, output_scale):
         m_ext_p, m0_p, p, p_inv = output_begin.cache
@@ -135,26 +125,29 @@ class _IsoIBM(_collections.AbstractExtrapolation):
         # The backward models remains preconditioned, because
         # we do backward passes in preconditioner-space.
         l_ext = p[:, None] * l_ext_p
-        m_bw = p[:, None] * m_bw_p
+        m_bw = p * m_bw_p
         l_bw = p[:, None] * l_bw_p
         g_bw = p[:, None] * g_bw_p * p_inv[None, :]
 
-        backward_noise = _vars.IsoNormalHiddenState(mean=m_bw, cov_sqrtm_lower=l_bw)
-        bw_model = _conds.IsoConditionalHiddenState(g_bw, noise=backward_noise)
-        extrapolated = _vars.IsoNormalHiddenState(mean=m_ext, cov_sqrtm_lower=l_ext)
-        return _vars.IsoStateSpaceVar(extrapolated, cache=None), bw_model
+        backward_noise = _vars.NormalHiddenState(mean=m_bw, cov_sqrtm_lower=l_bw)
+        bw_model = _conds.ConditionalHiddenState(g_bw, noise=backward_noise)
+        extrapolated = _vars.NormalHiddenState(mean=m_ext, cov_sqrtm_lower=l_ext)
+        return _vars.StateSpaceVar(extrapolated, cache=None), bw_model
 
-    # todo: should this be a classmethod in _conds.IsoConditional?
     def init_conditional(self, ssv_proto):
         op = self._init_backward_transition()
         noi = self._init_backward_noise(rv_proto=ssv_proto.hidden_state)
-        return _conds.IsoConditionalHiddenState(op, noise=noi)
+        return _conds.ConditionalHiddenState(op, noise=noi)
 
     def _init_backward_transition(self):
-        return jnp.eye(*self.a.shape)
+        k = self.num_derivatives + 1
+        return jnp.eye(k)
 
-    def _init_backward_noise(self, rv_proto):
-        return _vars.IsoNormalHiddenState(
-            mean=jnp.zeros_like(rv_proto.mean),
-            cov_sqrtm_lower=jnp.zeros_like(rv_proto.cov_sqrtm_lower),
-        )
+    @staticmethod
+    def _init_backward_noise(rv_proto):
+        mean = jnp.zeros_like(rv_proto.mean)
+        cov_sqrtm_lower = jnp.zeros_like(rv_proto.cov_sqrtm_lower)
+        return _vars.NormalHiddenState(mean, cov_sqrtm_lower)
+
+    def init_output_scale(self, output_scale):
+        return output_scale