+++ tests/test_equivalences/test_final_state_filter.py
@@ -6,7 +6,7 @@ import jax.numpy as jnp
 
 from probdiffeq import ivpsolve, ivpsolvers
 from probdiffeq.backend import testing
-from probdiffeq.ssm import recipes
+from probdiffeq.statespace import recipes
 from probdiffeq.strategies import filters, smoothers