+++ tests/test_ivpsolve/test_simulate_terminal_values.py
@@ -10,7 +10,7 @@ import jax.test_util
 
 from probdiffeq import controls, ivpsolve, ivpsolvers, taylor, test_util
 from probdiffeq.backend import testing
-from probdiffeq.ssm import recipes
+from probdiffeq.statespace import recipes
 from probdiffeq.strategies import filters, smoothers
 
 # Generate interesting test cases
@@ -30,7 +30,7 @@ class _SimulateTerminalValuesConfig(NamedTuple):
 @testing.case
 @testing.parametrize_with_cases("ode_problem", cases="..problem_cases", has_tag="nd")
 @testing.parametrize_with_cases("impl_fn", cases="..impl_cases", has_tag="nd")
-def case_setup_all_ssm_nd(ode_problem, impl_fn, solver_config):
+def case_setup_all_statespace_nd(ode_problem, impl_fn, solver_config):
     return _SimulateTerminalValuesConfig(
         ode_problem=ode_problem,
         solver_fn=ivpsolvers.MLESolver,
@@ -48,7 +48,7 @@ def case_setup_all_ssm_nd(ode_problem, impl_fn, solver_config):
     "ode_problem", cases="..problem_cases", has_tag="scalar"
 )
 @testing.parametrize_with_cases("impl_fn", cases="..impl_cases", has_tag="scalar")
-def case_setup_all_ssm_scalar(ode_problem, impl_fn, solver_config):
+def case_setup_all_statespace_scalar(ode_problem, impl_fn, solver_config):
     return _SimulateTerminalValuesConfig(
         ode_problem=ode_problem,
         solver_fn=ivpsolvers.MLESolver,