+++ tests/test_ivpsolve/test_solve_fixed_grid.py
@@ -8,8 +8,8 @@ import jax.test_util
 
 from probdiffeq import ivpsolve, ivpsolvers, test_util
 from probdiffeq.backend import testing
-from probdiffeq.ssm import recipes
-from probdiffeq.ssm.dense import corr as dense_corr
+from probdiffeq.statespace import recipes
+from probdiffeq.statespace.dense import corr as dense_corr
 from probdiffeq.strategies import filters, smoothers
 
 
@@ -25,7 +25,7 @@ class _SolveFixedGridConfig(NamedTuple):
 @testing.case
 @testing.parametrize_with_cases("ode_problem", cases="..problem_cases", has_tag="nd")
 @testing.parametrize_with_cases("impl_fn", cases="..impl_cases", has_tag="nd")
-def case_setup_all_ssm_nd(ode_problem, impl_fn, solver_config):
+def case_setup_all_statespace_nd(ode_problem, impl_fn, solver_config):
     return _SolveFixedGridConfig(
         ode_problem=ode_problem,
         solver_fn=ivpsolvers.MLESolver,
@@ -41,7 +41,7 @@ def case_setup_all_ssm_nd(ode_problem, impl_fn, solver_config):
     "ode_problem", cases="..problem_cases", has_tag="scalar"
 )
 @testing.parametrize_with_cases("impl_fn", cases="..impl_cases", has_tag="scalar")
-def case_setup_all_ssm_scalar(ode_problem, impl_fn, solver_config):
+def case_setup_all_statespace_scalar(ode_problem, impl_fn, solver_config):
     return _SolveFixedGridConfig(
         ode_problem=ode_problem,
         solver_fn=ivpsolvers.MLESolver,